Split Monument.js up into child components
