+++ client/src/components/Monument/Gallery/Gallery.scss
@@ -1,4 +1,4 @@
-@import "../../theme";
+@import "../../../theme";
 
 .images {
     display: flex;
