+++ client/src/components/Monument/Monument.scss
