+++ client/src/pages/MonumentPage/MonumentPage.scss
@@ -1,0 +1,147 @@
+@import "../../../node_modules/bootstrap/scss/functions";
+@import "../../../node_modules/bootstrap/scss/variables";
+@import "../../../node_modules/bootstrap/scss/mixins";
+
+.page-container {
+    min-width: 100%;
+    margin: 0;
+    display: flex;
+    flex-direction: column;
+
+    .column {
+        flex: 1 1 100%;
+        padding: 1rem 0;
+        &.main-column {
+            padding-top: 0;
+        }
+        &.related-monuments-column {
+            order: 3;
+            padding-bottom: 0;
+        }
+
+        > div {
+            margin-bottom: 2rem;
+
+            &:last-child {
+                margin-bottom: 0;
+            }
+        }
+    }
+
+    @include media-breakpoint-up(lg) {
+        flex-direction: row;
+        .column {
+            flex-basis: 35%;
+            padding: 0 1rem;
+            &.related-monuments-column {
+                order: unset;
+                padding-left: 0;
+                flex-basis: 30%;
+            }
+            &.visit-column {
+                padding-right: 0;
+            }
+        }
+    }
+}
+
+.main {
+
+    .fields .field {
+        margin-bottom: 1rem;
+    }
+
+    .image-container {
+        margin: 1rem 0;
+        background-color: rgb(30,30,30);
+        position: relative;
+        cursor: pointer;
+
+        .image-wrapper {
+            display: flex;
+            justify-content: center;
+        }
+
+        img, .overlay {
+            max-height: 300px;
+            width: auto;
+            max-width: 100%;
+        }
+
+        .overlay {
+            position: absolute;
+            background-color: rgba(0,0,0,0);
+            width: 100%;
+            height: 100%;
+            top: 0;
+            transition: background-color 0.1s ease-out;
+
+            .icon-wrapper {
+                display: flex;
+                align-items: center;
+                justify-content: center;
+                width: 100%;
+                height: 100%;
+                opacity: 0;
+                transition: opacity 0.1s ease-out;
+            }
+
+            i {
+                color: white;
+                font-size: 2rem;
+            }
+        }
+
+        &:hover {
+            .overlay {
+                background-color: rgba(0,0,0,0.5);
+            }
+            .icon-wrapper {
+                opacity: 1;
+            }
+        }
+    }
+}
+
+.modal-body {
+    img {
+        max-width: 100%;
+    }
+}
+
+.visit, .main {
+
+    .card-title {
+        margin-bottom: 0;
+        padding-bottom: 0.5rem;
+    }
+
+    .card-body {
+        margin-top: 0;
+        padding-top: 0;
+    }
+}
+
+.visit {
+    .card-title {
+        padding-bottom: 1rem;
+    }
+
+    iframe, .mapouter, .gmap_canvas {
+        width: 100%;
+        min-height: 400px;
+    }
+}
+
+.detail-list {
+    ul, ol {
+        margin-bottom: 0;
+    }
+    & > div {
+        margin-bottom: 0.25rem;
+    }
+}
+
+.detail-label {
+    font-weight: 500;
+}
